debt - remove unused snippet code
